First elements generation iter

+++ DataStructures/Geometry/CircleMaterial.cs
@@ -1,3 +1,3 @@
 ﻿namespace DataStructures.Geometry;
 
-public record struct CircleMaterial(int Degrees, double Material);
+public record struct CircleMaterial(double Degrees, double Material);
